Switch to the async API and object_store_ffi infra (#3)
Adds an async API to the Julia FFI.

Asynchronous calls are supported by allowing users to pass a condition
variable which is notified on events. The PR also refactors the way
objects from iceberg-rust are represented in the C-API. In particular,
we now provide an `IcebergScan` object which corresponds to a `Scan` in
iceberg-rust. Properties like batch size and data file concurrency, etc.
can be passed to the `IcebergScan`.

---------

Co-authored-by: Gerald Berger <gberger.work@gmail.com>